Introduced MapView interface with StaticMapView implementation
